[#IOPLT-1002] Rename OpenAPI introspection file name
