+++ apps/io-session-manager/src/controllers/session.ts
@@ -37,7 +37,7 @@ import { SuccessResponse } from "../generated/backend/SuccessResponse";
 import { log } from "../utils/logger";
 import { Concat, Union2Tuple, parseFilter } from "../utils/fields-filter";
 import { AssertionRef } from "../generated/lollipop-api/AssertionRef";
-import { UserIdentityWithTtl } from "../generated/external/UserIdentityWithTtl";
+import { UserIdentityWithTtl } from "../generated/introspection/UserIdentityWithTtl";
 
 // how many random bytes to generate for each session token
 export const SESSION_TOKEN_LENGTH_BYTES = 48;
